Rename content system module and update imports
